+++ src/lib/components/index/founding-members.svelte
@@ -1,5 +1,5 @@
 <script>
-	import Section from './section.svelte';
+	import Section from '../library/section.svelte';
 </script>
 
 <Section title="Founding members">
