+++ src/lib/components/index/learn-more.svelte
@@ -1,6 +1,6 @@
 <script lang="ts">
 	import { learnMoreItems } from '$lib/contents';
-	import Section from './section.svelte';
+	import Section from '../library/section.svelte';
 </script>
 
 <Section
