+++ src/routes/+page.svelte
@@ -1,8 +1,8 @@
 <script>
-	import FoundingMembers from '$lib/components/founding-members.svelte';
-	import Hero from '$lib/components/hero.svelte';
-	import LearnMore from '$lib/components/learn-more.svelte';
-	import Section from '$lib/components/section.svelte';
+	import FoundingMembers from '$lib/components/index/founding-members.svelte';
+	import Hero from '$lib/components/index/hero.svelte';
+	import LearnMore from '$lib/components/index/learn-more.svelte';
+	import Section from '$lib/components/library/section.svelte';
 
 	import GradientLinkButton from '$lib/components/library/gradient-link-button.svelte';
 </script>
@@ -12,15 +12,15 @@
 <Section
 	title="What is developer experience?"
 	text="DevX Conf is a space for creators where we can start from first principles, collaborate, listen, discuss and declutter our workflows, toolchains and minds. To improve developer experience. To make us feel again like everything is at our fingertips. We want to foster an ecosystem that sparks dialogue and establishes best practices and frameworks across the areas of coding, testing, building, deploying, collaborating on and running software."
-	textClassNames="max-w-[558px]"
+	textClassNames="max-w-xl"
 	dots={true}
 />
 
 <Section
-	title="What is this community? And how to join?"
+	title="What is this community?"
 	text="DevX Conf is a space for creators where we can start from first principles, collaborate, listen, discuss and declutter our workflows, toolchains and minds. To improve developer experience. To make us feel again like everything is at our fingertips. We want to foster an ecosystem that sparks dialogue and establishes best practices and frameworks across the areas of coding, testing, building, deploying, collaborating on and running software."
 	titleClassNames="max-w-xl mx-auto"
-	textClassNames="max-w-[558px]"
+	textClassNames="max-w-xl"
 >
 	<div class="mt-x-large flex justify-center">
 		<GradientLinkButton href="">Join the community</GradientLinkButton>
